missing assignment condition, typos correction
In my case 'climb' was not part of a packet I got, so an exception was raised and script wasn't succeeding. Now it is, at least in my case.
Here is a list of elements that have to or not be in a packet: http://www.catb.org/gpsd/gpsd_json.html
According to this packet can have or have not most of the elements, so I added appropriate conditions on them all.